pyproject.toml
[tool.poetry]
name = "lifxlightlan"
version = "0.1.0"
description = ""
authors = ["praneethbluecats98 <73488628+praneethbluecats98@users.noreply.github.com>"]

[tool.poetry.dependencies]
python = "^3.10"
lifxlan = "^1.2.7"
paho-mqtt = "^1.6.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
